simpleapp/requirements/production.txt
-r ./base.txt

gunicorn==20.0.4
django-mysql==3.7.0
mysqlclient==1.4.6
django-redis==4.12.1
redis==3.5.3